Move lorebook signal detection into its own module
Inlining the tag/category signal tables pushed normalization.rs to 582 lines, over the 450-line cap enforced by check:architecture (scripts/check-rust-structure.mjs). Extract entry_signal_text, detect_entry_tag, detect_category, and the signal tables into a sibling lorebook_signals module, bringing normalization.rs back to 450 and keeping import normalization isolated and reviewable. Also drop a redundant second lowercasing pass in detect_category (entry blobs are already lowercased).
